OF-1988: Introduce pluggable 'provider' mechanism for PubSub perisistency.
This commit replaces the static methods that store PubSub data in the database with a provider for such data-peristency functionality. This is not unlike the provider mechanism that is available for things like Users, Groups and Admins.

Additionally, bug fixes / improvements have been applied to earlier work. Notably, the identification of Node instances now no longer is based solely on the Node ID field (which is only unique in context of one service). Instead, an "unique identifier" is used, that combines server ID and node ID.