OF-1021: Verify content type of uploaded plugins
Commit 9c62dbf599f266bde214c5a68a004708edc7da48 adds code that removes uploaded files that cannot be parsed as JAR files.

To further guard against malicous files being uploaded, this commit adds functionality to allow to verify the content type, as specified by the browser, of the uploaded file.

As a potential attacker is likely to be able to modify the reported content type. The added security value of this change is therefor not very signification. By default, this functionality is therefor disabled, to prevent valid use cased from being stopped by this.

The new functionality is controlled by two new properties:
- plugins.upload.content-type-check.enabled A boolean value that enables or disables the check (defaults to false).
- plugins.upload.content-type-check.expected-value Text value that is the expected content type (defaults to application/x-java-archive).