OF-1883: Admin console should list client sessions that are detached.
OF-590 introduces a connection validity check to the client sessions that are to be listed on the admin console. This check fails for sessions that are detached, which caused them to go unlisted.

This commit removes the check that was introduced in OF-590. I feel confident that the problems that guarded for are fixed by OF-1700.

Additionally, some (debug) log statements were added / enriched with session data.