OF-353: Allow baseDN to have spaces
The baseDN value is used to compose the connect host(s) URL. This value is a space-separated string of URLs. When a baseDN contains a space character, this character must be escaped, to prevent it from being interpreted as a second connect host.

This commit adds URL encoding to the baseDN when it's used to create the connect host URL.