OF-1944: Clear caches after changing LDAP settings.
When LDAP profile settings are changed (for example, to change the VCard mapping), then old data can still linger in the caches. To the unexpected user, it'll appear as if changes have not been applied.

To help changes be effectuated immediately, relevant caches should be cleared after updating the LDAP profile settings.