OF-2014: Update OnConnectionClose callback object when re-initializing connection.
When a connection gets a new owner (Session), make sure that the old owner that was registered with the callback function gets replaced by the new owner. Without this, the callback will be invoked on the old (non-existing) session. This, among others, prevents Stream Management from working properly after a stream has been resumed for the first time.