OF-2020: Don't recreate pepservice in process of deleting it.
When a PEP Service is destroyed, all of its nodes are destroyed as well.

The current implementation first deletes the PEP service itself, and then iterates over all of its nodes to delete them.

When deleting a node, the internal code tries to obtain its PEP service. As that was just deleted, it's auto-created again (or, alternatively, an exception occurs).

Openfire should first delete the nodes and only then remove the PEP service to avoid this issue.