OF-2030: (FMUC) Two-way join direction, separation of concerns
Adding support for both inbound as well as outbound joins.

Separating concerns for FMUC handling from regular FMUC implementation.

Added implementation of processing of most (but not all) stanzas that are exchanged through FMUC.

This commit is the result of a few weeks of work and should probably have been split up in multiple distinct commits.