OF-2040: Remove entity capabilities as soon as a new 'ver' arrives.
Prior to this change, the 'old' entity capabilities for a user were retained as long as it took to resolve the updated 'ver' value. This leads to timing-sensitive bugs, where Openfire, briefly, exposes the 'wrong' capabilities.

This commit removes the old capabilities immediately, storing them (temporarily) to be able to diff them with the update (which is used in the new event listeners).