OF-2146: Undo work around self-presence race-condition introduced in OF-2030
The work-around depends on the sender of the presence to also be the subject of the presence, which cannot be relied on. For example, a presence stanza that announces that someone has been granted the role of moderator is sent from the room address, not the end-user entity that did the granting.