OF-2206: When closing BOSH session, send 'terminate'
When a BOSH session gets terminated, the client should be told that not only the request is 'done' (eg: gets responded to) but also the session that the connection is part of is terminated. Without this, there is no way for the client to determine that this is the case, other than receiving an error on the next request where the client attempts to re-use the session (sends a request with the old SID value).