OF-2212: Allow admin console's cert to be auto-updated
Openfire detects changes to its keystores, and reloads the various services that use them. This, for example, allows the CertificateManager plugin to update Openfire's security certificates dynamically.

One element that does not automatically get updated after keystore changes is the webserver that is serving the admin console (which also can serve other web content). This is a conscious choice, as the admin console can be used to update the keystores. It would be very undesirable to have the admin user be logged out as a result of applying a change to the keystores. To work around this issue, the admin console remains unchanged when a keystore change is detected (although a warning will be displayed, telling an admin to restart things).

This commit ensures that the admin console is automatically updated when certificate changes are detected, unless that change happened through the admin console, by briefly pausing the auto-updating feature.