Fixes for OF-2246, OF-2249 and OF-2250
This commit started off as an implementation of a new feature that allows backup delivery to be disabled (OF-2246). While implementing that, two bugs were discovered with how backup delivery is applied for BOSH and multiplexers (OF-2249, OF-2250) for which the fix has been included in this commit.

A slight refactoring of the BOSH connection implementation was added, to bring those more in-line with Connection implementations of other connection types.