OF-2257: Replace usage of TaskEngine prior to server initialization
Prioer to Openfire being initialized, TaskEngine should not be used. This prevents issues with TaskEngine using properties that are yet to be initialized. Although TaskManager does not currently do that in its present state, it will, when support for JMX is added (OF-2257).

This commit replaces the usage of TaskEngine with a generic Timer.