OF-2254: Admin console to show if system property was set by user
Apart from distinghishing between a property value that is, or is not equal to the default value of the property, it is convenient to be able to tell if a property is equal to the default value (this helps to guard against property values for which the default is changed between Openfire releases). This commit adds a column that can be used to determine if the property value has been explicitly set (instead of it just being a default).